Add Post and Comments component
